+++ sdk/ml/azure-ai-ml/azure/ai/ml/operations/_registry_operations.py
@@ -137,14 +137,16 @@ class RegistryOperations:
 
 
     # @monitor_with_activity(logger, "Registry.Delete", ActivityType.PUBLICAPI)
-    def delete(self, *, name: str, **kwargs: Dict) -> None:
+    def delete(self, *, name: str, **kwargs: Dict) -> LROPoller[Registry]:
         """Delete a registry. Returns nothing on a successful operation.
 
         :param name: Name of the registry
         :type name: str
+        :return: A poller to track the operation status.
+        :rtype: LROPoller
         """
         resource_group = kwargs.get("resource_group") or self._resource_group_name
-        return self._operation.delete(
+        return self._operation.begin_delete(
             resource_group_name=resource_group,
             registry_name=name,
             **self._init_kwargs,