+++ sdk/ml/azure-ai-ml/CHANGELOG.md
@@ -11,7 +11,7 @@
 
 ### Bugs Fixed
 - MLClient.from_config can now find the default config.json on Compute Instance when running sample notebooks.
-
+- Make registry delete operation return an LROPoller.
 ### Other Changes
 
 ## 1.0.0 (2022-10-10)